Fix #162
sheet - use only in R to denote an argument that can either be a name or index
sheet_index - use in R and C++ to denote an integer representing the sheet index, e.g. 2
sheet_name - use in R and C++ to denote a string representing the sheet name, e.g. "Sheet1"
sheet_file - use in R and C++ to denote the often throwaway XML file